refactor: move existing rest controllers to v0.1 folder
Move existing rest controllers for W3C-API to v0.1 as to be in sync with v1.0 folder containing OIDC/W3C APIs
