+++ cmd/vc-rest/startcmd/start.go
@@ -27,15 +27,15 @@ import (
 
 	"github.com/trustbloc/vcs/api/spec"
 	"github.com/trustbloc/vcs/cmd/common"
-	restholder "github.com/trustbloc/vcs/pkg/restapi/holder"
-	holderops "github.com/trustbloc/vcs/pkg/restapi/holder/operation"
-	restissuer "github.com/trustbloc/vcs/pkg/restapi/issuer"
-	issuerops "github.com/trustbloc/vcs/pkg/restapi/issuer/operation"
+	restholder "github.com/trustbloc/vcs/pkg/restapi/v0.1/holder"
+	holderops "github.com/trustbloc/vcs/pkg/restapi/v0.1/holder/operation"
+	restissuer "github.com/trustbloc/vcs/pkg/restapi/v0.1/issuer"
+	issuerops "github.com/trustbloc/vcs/pkg/restapi/v0.1/issuer/operation"
+	restverifier "github.com/trustbloc/vcs/pkg/restapi/v0.1/verifier"
+	verifierops "github.com/trustbloc/vcs/pkg/restapi/v0.1/verifier/operation"
 	"github.com/trustbloc/vcs/pkg/restapi/v1/healthcheck"
 	issuerv1 "github.com/trustbloc/vcs/pkg/restapi/v1/issuer"
 	verifierv1 "github.com/trustbloc/vcs/pkg/restapi/v1/verifier"
-	restverifier "github.com/trustbloc/vcs/pkg/restapi/verifier"
-	verifierops "github.com/trustbloc/vcs/pkg/restapi/verifier/operation"
 	"github.com/trustbloc/vcs/pkg/storage/mongodb"
 	"github.com/trustbloc/vcs/pkg/storage/mongodb/verifierstore"
 	verifiersvc "github.com/trustbloc/vcs/pkg/verifier"
