+++ pkg/restapi/v0.1/holder/controller.go
@@ -4,22 +4,22 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package issuer
+package holder
 
 import (
-	"github.com/trustbloc/vcs/pkg/restapi/issuer/operation"
+	"github.com/trustbloc/vcs/pkg/restapi/v0.1/holder/operation"
 )
 
 // New returns new controller instance.
 func New(config *operation.Config) (*Controller, error) {
 	var allHandlers []operation.Handler
 
-	vcService, err := operation.New(config)
+	holderService, err := operation.New(config)
 	if err != nil {
 		return nil, err
 	}
 
-	handlers := vcService.GetRESTHandlers()
+	handlers := holderService.GetRESTHandlers()
 
 	allHandlers = append(allHandlers, handlers...)
