+++ pkg/restapi/v0.1/holder/controller_test.go
@@ -10,6 +10,8 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/trustbloc/vcs/pkg/restapi/v0.1/holder/operation"
+
 	"github.com/trustbloc/vcs/pkg/storage/ariesprovider"
 
 	ariesmemstorage "github.com/hyperledger/aries-framework-go/component/storageutil/mem"
@@ -17,8 +19,6 @@ import (
 	ariesmockstorage "github.com/hyperledger/aries-framework-go/pkg/mock/storage"
 	vdrmock "github.com/hyperledger/aries-framework-go/pkg/mock/vdr"
 	"github.com/stretchr/testify/require"
-
-	"github.com/trustbloc/vcs/pkg/restapi/holder/operation"
 )
 
 func TestController_New(t *testing.T) {
