+++ pkg/restapi/v0.1/internal/common/did/common.go
