+++ pkg/restapi/v0.1/internal/common/vcutil/vcutil.go
