+++ pkg/restapi/v0.1/verifier/operation/models.go
