+++ pkg/restapi/v0.1/verifier/operation/operations_test.go
